feat(milk): move constants to dedicated directory
