+++ milk/src/connections/sentry.ts
@@ -1,7 +1,8 @@
 import * as Sentry from '@sentry/node'
 import {Logger} from "@beemobot/common";
 // ^ This needs to be updated; Probably @beemobot/cafe
-import {TAG} from "../index.js";
+
+import {TAG} from "../constants/logging.js";
 export function initializeSentry() {
     if (process.env.SENTRY_DSN == null) {
         Logger.warn(TAG, 'Sentry is not configured, we recommend configuring Sentry to catch issues properly.')
