+++ milk/src/constants/logging.ts
@@ -1,0 +1,1 @@
+export const TAG = "Milk"
