+++ milk/src/constants/raid_management_kafka.ts
@@ -1,0 +1,3 @@
+export const RAID_MANAGEMENT_CLIENT_TOPIC = "raid-management"
+export const RAID_MANAGEMENT_BATCH_INSERT_KEY = "batch-insert-raid-users"
+export const RAID_MANAGEMENT_CONCLUDE_RAID = "conclude-raid"
