+++ milk/src/constants/time.ts
@@ -1,0 +1,1 @@
+export const TEN_MINUTES = 60 * 10
