+++ milk/src/hooks/log_hook.ts
@@ -1,7 +1,7 @@
 import {Logger} from "@beemobot/common";
 // ^ This needs to be updated; Probably @beemobot/cafe
-import {TAG} from "../index.js";
 import {FastifyInstance} from "fastify";
+import {TAG} from "../constants/logging.js";
 
 export default async (fastify: FastifyInstance) => {
     fastify.addHook(
