+++ milk/src/index.ts
@@ -1,17 +1,14 @@
-import {PrismaClient} from "@prisma/client";
 import dotenv from 'dotenv'
 import {initializeSentry, logIssue} from "./connections/sentry.js";
 import {initializeKafka} from "./connections/kafka.js";
-import {initializePrisma} from "./connections/prisma.js";
+import {initializePrisma, prisma} from "./connections/prisma.js";
 import {Logger} from "@beemobot/common";
 // ^ This needs to be updated; Probably @beemobot/cafe
 import {initializeFastify} from "./connections/fastify.js";
 import * as Sentry from '@sentry/node'
+import {TAG} from "./constants/logging.js";
 
 dotenv.config()
-
-export let prisma = new PrismaClient()
-export const TAG = "Milk"
 
 async function main() {
     initializeSentry()
